Code spliting with the React Context API
