+++ package.json
@@ -120,6 +120,7 @@
     "@types/mdx": "^2.0.5",
     "@types/node": "^18.16.18",
     "@types/react-inner-image-zoom": "^3.0.0",
+    "@types/react-timeago": "^4.1.3",
     "@types/uuid": "^9.0.2",
     "@vitejs/plugin-react": "^3.1.0",
     "axios": "^1.4.0",
