+++ src/mangadex/pages/chapter/ReadingOptionDrawer/index.tsx
@@ -2,27 +2,28 @@ import React from "react";
 import { Chapter } from "@mangadex/api/structures/Chapter";
 import * as Chakra from "@chakra-ui/react";
 import useChapterReadingDrawer from "@mangadex/resources/hooks/fullscreenOption";
-import useChapterReadingModeOption from "../ChapterReadingMode/useChapterReadingModeOption";
-import { ReadingMode } from "@mangadex/api/internal/UserOptions/ReadingMode";
 
-const DirectionSelection = React.lazy(() => import("../DirectionSelection"));
-
-
-
+const PageSelection = React.lazy(() => import("./PageSelection"));
+const ChapterReadingModeOption = React.lazy(() => import("./Chapter_Reading_Mode"));
+const Direction = React.lazy(() => import("./Direction"));
+const ImageWidthController = React.lazy(() => import("./ImageWidthController"));
+const NavigationModal = React.lazy(() => import("./NavigationModal"));
+const OptionOverlay = React.lazy(() => import("./OptionOverlay"));
+const Navigation = React.lazy(() => import("./Navigation"));
 
 const ReadingDrawerContext = React.createContext<Chapter | undefined>(undefined);
 
-export function useReadingDraweContext(){
+export function useReadingDraweContext() {
     const data = React.useContext(ReadingDrawerContext);
-    if(data == undefined){
+    if (data == undefined) {
         throw new Error("The Reading Drawer Context Provider is not implemented");
     }
     return data;
 }
 
-export function ReadingDrawerContextProvider({ value , children } : React.PropsWithChildren<{
-    value : Chapter
-}>){
+export function ReadingDrawerContextProvider({ value, children }: React.PropsWithChildren<{
+    value: Chapter
+}>) {
     return (
         <ReadingDrawerContext.Provider value={value}>
             {children}
@@ -34,42 +35,17 @@ export default function ReadingDrawer(props: {
     chapter: Chapter,
 }) {
     const { query, changeOption } = useChapterReadingDrawer();
-    
-    function Direction() {
-        const reading_mode = useChapterReadingModeOption();
-        if (reading_mode.query.data != ReadingMode.LongStrip) {
-            return (
-                <Chakra.HStack>
-                    <Chakra.Text as={"span"}>Direction : </Chakra.Text>
-                    <React.Suspense>
-                        <DirectionSelection />
-                    </React.Suspense>
-                </Chakra.HStack>
-            );
-        } else {
-            return (
-                <React.Fragment />
-            );
-        }
-    }
+
+
     return (
-        <React.Fragment>
+        <ReadingDrawerContextProvider value={props.chapter}>
             <Chakra.Drawer
                 isOpen={query.data ?? false}
                 onClose={() => {
                     changeOption(false);
                 }}
             >
-                {
-                    isOverlay == true ? (
-                        <Chakra.DrawerOverlay
-                            zIndex={"100"}
-                        />
-                    ) : (
-                        <React.Fragment/>
-                    )
-                }
-                <Chakra.DrawerContent 
+                <Chakra.DrawerContent
                     zIndex={"100"}
                     fontFamily={"inherit"}
                 >
@@ -79,17 +55,27 @@ export default function ReadingDrawer(props: {
                     </Chakra.DrawerHeader>
                     <Chakra.DrawerBody>
                         <Chakra.VStack alignItems={"intial"} spacing={"10px"}>
-                            <Navigation />
-                            <OptionOverlay />
-                            <ImageWidthController />
-                            <ChapterReadingModeOption />
-                            <Direction />
-                            <NavigationModal />
-                            <PageSelection />
+                            <React.Suspense
+                                fallback={
+                                    <Chakra.Alert variant={"left-accent"} status="loading">
+                                        <Chakra.AlertIcon />
+                                        <Chakra.AlertTitle>Loading componnents...</Chakra.AlertTitle>
+                                        <Chakra.AlertDescription>Yes, everything is lazy loaded now</Chakra.AlertDescription>
+                                    </Chakra.Alert>
+                                }
+                            >
+                                <Navigation />
+                                <OptionOverlay />
+                                <ImageWidthController />
+                                <ChapterReadingModeOption />
+                                <Direction />
+                                <NavigationModal />
+                                <PageSelection />
+                            </React.Suspense>
                         </Chakra.VStack>
                     </Chakra.DrawerBody>
                 </Chakra.DrawerContent>
             </Chakra.Drawer>
-        </React.Fragment>
+        </ReadingDrawerContextProvider>
     );
 }
